Renamed files to better match new type names
